Extract TinyMCE imports and config into separate file
